# bitset: Improve performance of count function