The `count()` function of the `bitset` container used two distinct reductions for computing the number of set bits. The first one considers all bits in the first `n - 1` block whereas the second one works on each of the remaining bits in the `n`-th block, i.e. the last block. Since the reduction is memory-bound anyways, the second reduction introduces a large overhead while the first one still has room for additional compute operations. Merge both reductions into a single one that computes a bit mask for each block on the fly before counting the bits per block. This leads to a decent performance improvement for `count()` and all related functions.

Before:
```
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
stdgpu_bitset_count/1000            0.050 ms        0.051 ms        13822
stdgpu_bitset_count/1000000         0.056 ms        0.056 ms        12476
stdgpu_bitset_count/1000000000      0.733 ms        0.733 ms          956
stdgpu_bitset_all/1000              0.050 ms        0.050 ms        13879
stdgpu_bitset_all/1000000           0.056 ms        0.056 ms        12538
stdgpu_bitset_all/1000000000        0.736 ms        0.734 ms          954
stdgpu_bitset_any/1000              0.050 ms        0.050 ms        13841
stdgpu_bitset_any/1000000           0.056 ms        0.056 ms        12165
stdgpu_bitset_any/1000000000        0.734 ms        0.734 ms          954
stdgpu_bitset_none/1000             0.050 ms        0.051 ms        13811
stdgpu_bitset_none/1000000          0.056 ms        0.056 ms        12428
stdgpu_bitset_none/1000000000       0.735 ms        0.734 ms          954
```

After:
```
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
stdgpu_bitset_count/1000            0.026 ms        0.026 ms        26891
stdgpu_bitset_count/1000000         0.031 ms        0.031 ms        22745
stdgpu_bitset_count/1000000000      0.580 ms        0.580 ms         1208
stdgpu_bitset_all/1000              0.026 ms        0.026 ms        26856
stdgpu_bitset_all/1000000           0.031 ms        0.031 ms        22795
stdgpu_bitset_all/1000000000        0.580 ms        0.580 ms         1207
stdgpu_bitset_any/1000              0.026 ms        0.026 ms        26940
stdgpu_bitset_any/1000000           0.031 ms        0.031 ms        22814
stdgpu_bitset_any/1000000000        0.580 ms        0.580 ms         1207
stdgpu_bitset_none/1000             0.026 ms        0.026 ms        26945
stdgpu_bitset_none/1000000          0.032 ms        0.032 ms        22725
stdgpu_bitset_none/1000000000       0.581 ms        0.580 ms         1185
```